requirements.txt
streamlit
fastapi
uvicorn
requests
python-docx
fpdf2
pypdf
scikit-learn
numpy
passlib[bcrypt]
python-jose
python-dotenv
